stable refactor
