+++ packages/docusaurus-theme-classic/src/theme/DocPage/Layout/Aside/ExpandButton.tsx
@@ -1,0 +1,40 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import React from 'react';
+import IconArrow from '@theme/IconArrow';
+import {translate} from '@docusaurus/Translate';
+import type {Props} from '@theme/DocPage/Layout/Aside/ExpandButton';
+
+import styles from '../index.module.css';
+
+export default function DocPageLayoutAsideExpandButton({
+  toggleSidebar,
+}: Props): JSX.Element {
+  return (
+    <div
+      className={styles.collapsedDocSidebar}
+      title={translate({
+        id: 'theme.docs.sidebar.expandButtonTitle',
+        message: 'Expand sidebar',
+        description:
+          'The ARIA label and title attribute for expand button of doc sidebar',
+      })}
+      aria-label={translate({
+        id: 'theme.docs.sidebar.expandButtonAriaLabel',
+        message: 'Expand sidebar',
+        description:
+          'The ARIA label and title attribute for expand button of doc sidebar',
+      })}
+      tabIndex={0}
+      role="button"
+      onKeyDown={toggleSidebar}
+      onClick={toggleSidebar}>
+      <IconArrow className={styles.expandSidebarButtonIcon} />
+    </div>
+  );
+}
