+++ packages/docusaurus-theme-classic/src/theme/DocPage/Layout/Aside/index.tsx
@@ -7,40 +7,14 @@
 
 import React, {type ReactNode, useState, useCallback} from 'react';
 import DocSidebar from '@theme/DocSidebar';
-import IconArrow from '@theme/IconArrow';
-import {translate} from '@docusaurus/Translate';
 import {useLocation} from '@docusaurus/router';
 import type {Props} from '@theme/DocPage/Layout/Aside';
+import ExpandButton from '@theme/DocPage/Layout/Aside/ExpandButton';
 
 import clsx from 'clsx';
-import styles from './styles.module.css';
+import styles from './index.module.css';
 
 import {ThemeClassNames, useDocsSidebar} from '@docusaurus/theme-common';
-
-function SidebarExpandButton({toggleSidebar}: {toggleSidebar: () => void}) {
-  return (
-    <div
-      className={styles.collapsedDocSidebar}
-      title={translate({
-        id: 'theme.docs.sidebar.expandButtonTitle',
-        message: 'Expand sidebar',
-        description:
-          'The ARIA label and title attribute for expand button of doc sidebar',
-      })}
-      aria-label={translate({
-        id: 'theme.docs.sidebar.expandButtonAriaLabel',
-        message: 'Expand sidebar',
-        description:
-          'The ARIA label and title attribute for expand button of doc sidebar',
-      })}
-      tabIndex={0}
-      role="button"
-      onKeyDown={toggleSidebar}
-      onClick={toggleSidebar}>
-      <IconArrow className={styles.expandSidebarButtonIcon} />
-    </div>
-  );
-}
 
 // Reset sidebar state when sidebar changes
 // Use React key to unmount/remount the children
@@ -94,7 +68,7 @@ export default function DocPageLayoutAside({
         />
       </ResetOnSidebarChange>
 
-      {hiddenSidebar && <SidebarExpandButton toggleSidebar={toggleSidebar} />}
+      {hiddenSidebar && <ExpandButton toggleSidebar={toggleSidebar} />}
     </aside>
   );
 }
